test(file-utils.spec.js openid-client-helper.spec.js): separate unit and integration test
ref #206
