+++ test/openid-client-helper.test.js
@@ -12,8 +12,6 @@ const passportConfigAuthorizedResponse = config.get('passportConfigAuthorizedRes
 
 describe('Test OpenID Client Helper', () => {
   const testProvider = passportConfigAuthorizedResponse.providers.find(p => p.id === 'oidccedev6privatejwt')
-  let kid = null
-  const jwksFilePath = `../server/jwks/${testProvider.id}.json`
 
   describe('generateJWKS test', () => {
     const generateJWKS = rewiredOpenIDClientHelper.__get__('generateJWKS')
@@ -24,24 +22,6 @@ describe('Test OpenID Client Helper', () => {
 
     it('should be function', () => {
       assert.isFunction(generateJWKS, 'generateJWKS is not a function')
-    })
-
-    it('should generate jwks for provider in jwks folder', async () => {
-      await generateJWKS(testProvider)
-      assert.exists(jwksFilePath, `${jwksFilePath} file not found`)
-    })
-
-    it('make sure jwks has keys and kid', () => {
-      const jwks = require(jwksFilePath)
-      assert.isArray(jwks.keys, 'keys not found in jwks')
-      kid = jwks.keys[0].kid
-      assert.exists(kid, 'kid not found in jwks')
-    })
-
-    it('make sure generateJWKS not regenerating jwks again and rewrite existing jwks data', async () => {
-      await generateJWKS(testProvider)
-      const jwks = require(jwksFilePath)
-      assert.equal(kid, jwks.keys[0].kid, `${kid} is not matching with ${jwks.keys[0].kid}`)
     })
   })
